fix: scope component registry to compiler instance
Adding the registry as static instance to the component prevents garbage
collection of all compile trees when reusing the compiler in process.
Scoping it to the compiler instance make more sense, since we only need
to cache components for a compiler instance and it as well allows
garbage collection to do its work after disposing the compiler.